add first implementation of R1: BVH construction and traversal

+++ assignment1/src/base/Bvh.cpp
@@ -10,10 +10,10 @@ namespace FW {
 Bvh::Bvh() {
 }
 
-Bvh::Bvh(SplitMode splitmode, size_t start, size_t end) {
+Bvh::Bvh(SplitMode splitmode, uint32_t start, uint32_t end) {
     mode_ = splitmode;
     rootNode_ = std::make_unique<BvhNode>(start, end);
-    for (int i = start; i <= end; ++i) {
+    for (uint32_t i = start; i <= end; ++i) {
         indices_.push_back(i);
     }
 }
